share/requirements.txt
redis==5.0.7
mock==5.1.0

django==5.0.7
djangorestframework==3.15.2
django-oauth-toolkit==2.4.0
django-environ==0.11.2
django-cors-headers==4.4.0